backend/requirements.txt
django==4.2.1
djangorestframework==3.14.0
djoser==2.2.0
gunicorn==20.1.0
django-filter==23.2
Pillow==9.5.0
psycopg2_binary==2.9.6
